requirements.txt
requests
beautifulsoup4
openai
supabase
azure-ai-inference